Profile CRUD javítása, hibák logolása, ProfileActivity elrendezésének javítása

+++ app/src/main/java/com/example/playfulmath/ProfileActivity.java
@@ -17,6 +17,7 @@ import com.google.android.gms.tasks.Task;
 import com.google.firebase.FirebaseApp;
 import com.google.firebase.auth.FirebaseAuth;
 import com.google.firebase.auth.FirebaseUser;
+import com.google.firebase.auth.UserProfileChangeRequest;
 import com.google.firebase.database.DataSnapshot;
 import com.google.firebase.database.DatabaseError;
 import com.google.firebase.database.DatabaseReference;
@@ -53,32 +54,6 @@ public class ProfileActivity extends AppCompatActivity implements View.OnClickLi
         FirebaseApp.initializeApp(this);
     }
 
-    private void loadUserData() {
-        String userId = FirebaseAuth.getInstance().getCurrentUser().getUid();
-
-        DatabaseReference profileDb = FirebaseDatabase.getInstance().getReference("users").child(userId);
-
-        profileDb.addValueEventListener(new ValueEventListener() {
-            @Override
-            public void onDataChange(DataSnapshot dataSnapshot) {
-                if (dataSnapshot.exists()) {
-                    String username = dataSnapshot.child("username").getValue(String.class);
-                    String email = dataSnapshot.child("email").getValue(String.class);
-                    String password = dataSnapshot.child("password").getValue(String.class);
-                    String confirmPassword = dataSnapshot.child("password").getValue(String.class);
-
-                    profileUsernameEditText.setText(username);
-                    profileEmailEditText.setText(email);
-                    profilePasswordEditText.setText(password);
-                    profileConfirmPasswordEditText.setText(confirmPassword);
-                }
-            }
-            @Override
-            public void onCancelled(DatabaseError databaseError) {
-
-            }
-        });
-    }
     @Override
     public void onClick(View v) {
         Intent i;
@@ -105,11 +80,25 @@ public class ProfileActivity extends AppCompatActivity implements View.OnClickLi
                 } else if (newProfileEmail.isEmpty() || !newProfileEmail.contains("@")) {
                     showError(profileEmailEditText, "Az e-mail cím érvénytelen.");
                 } else {
+                    user.updateEmail(newProfileEmail)   //FirebaseAuth frissítése
+                            .addOnCompleteListener(new OnCompleteListener<Void>() {
+                                @Override
+                                public void onComplete(@NonNull Task<Void> task) {
+                                    if (task.isSuccessful()) {
+                                        // Akkor adatok módosítása Firebase Realtime Database-ban
+                                        profileDb.child("username").setValue(newProfileUsername);
+                                        profileDb.child("email").setValue(newProfileEmail);
 
-                    // Adatok módosítása firebaseban
-                    profileDb.child("username").setValue(newProfileUsername);
-                    profileDb.child("email").setValue(newProfileEmail);
-                    Toast.makeText(ProfileActivity.this, "Sikeres adat módosítás!", Toast.LENGTH_SHORT).show();
+                                        Toast.makeText(ProfileActivity.this, "Sikeres adat módosítás!", Toast.LENGTH_SHORT).show();
+                                    } else {
+                                        Exception exception = task.getException();
+                                        if (exception != null) {
+                                            Log.e("ProfileActivity", "Hiba történt az e-mail cím frissítése során: " + exception.getMessage());
+                                            showError(profileEmailEditText, "Sikertelen e-mail cím módosítás!");
+                                        }
+                                    }
+                                }
+                            });
                 }
                 break;
             case R.id.changePasswordButton:
@@ -125,32 +114,35 @@ public class ProfileActivity extends AppCompatActivity implements View.OnClickLi
                                 public void onComplete(@NonNull Task<Void> task) {
                                     if (task.isSuccessful()) {
                                         Toast.makeText(ProfileActivity.this, "Sikeres jelszó módosítás!", Toast.LENGTH_SHORT).show();
-                                        //TODO  UI frissítése
                                     } else {
-                                        Toast.makeText(ProfileActivity.this, "Sikertelen jelszó módosítás!", Toast.LENGTH_SHORT).show();
+                                        Exception exception = task.getException();
+                                        if (exception != null) {
+                                            Log.e("ProfileActivity", "Hiba történt a jelszó frissítése során: " + exception.getMessage());
+                                            showError(profilePasswordEditText, "Sikertelen jelszó módosítás!");
+                                        }
                                     }
                                 }
                             });
                 }
                 break;
             case R.id.deleteButton:
+                deleteUserDataFromDatabase();
                 user.delete()
                         .addOnCompleteListener(new OnCompleteListener<Void>() {
                             @Override
                             public void onComplete(@NonNull Task<Void> task) {
                                 if (task.isSuccessful()) {
                                     Toast.makeText(ProfileActivity.this, "Sikeres fiók törlés!", Toast.LENGTH_SHORT).show();
-                                    if (user != null) {
-                                        //TODO valamiért a reltime-ből nem törlődik
-                                        deleteUserDataFromDatabase();
-                                    } else {
-                                        Log.e("DeleteUserData", "Felhasználó nincs bejelentkezve.");
-                                    }
+
                                     Intent intent = new Intent(ProfileActivity.this,LoginActivity.class);
                                     startActivity(intent);
                                     finish();
                                 } else {
-                                    Toast.makeText(ProfileActivity.this, "Sikertelen fiók törlés!", Toast.LENGTH_SHORT).show();
+                                    Exception exception = task.getException();
+                                    if (exception != null) {
+                                        Log.e("ProfileActivity", "Hiba történt a profil törlése során: " + exception.getMessage());
+                                        Toast.makeText(ProfileActivity.this, "Sikertelen fiók törlés!", Toast.LENGTH_SHORT).show();
+                                    }
                                 }
                             }
                         });
@@ -163,27 +155,55 @@ public class ProfileActivity extends AppCompatActivity implements View.OnClickLi
         input.requestFocus();
     }
 
-    private void deleteUserDataFromDatabase() {
+
+    private void loadUserData() {
         String userId = FirebaseAuth.getInstance().getCurrentUser().getUid();
+
         DatabaseReference profileDb = FirebaseDatabase.getInstance().getReference("users").child(userId);
 
-        profileDb.removeValue()
-                .addOnCompleteListener(new OnCompleteListener<Void>() {
-                    @Override
-                    public void onComplete(@NonNull Task<Void> task) {
-                        if (task.isSuccessful()) {
+        profileDb.addValueEventListener(new ValueEventListener() {
+            @Override
+            public void onDataChange(DataSnapshot dataSnapshot) {
+                if (dataSnapshot.exists()) {
+                    String username = dataSnapshot.child("username").getValue(String.class);
+                    String email = dataSnapshot.child("email").getValue(String.class);
+                    String password = dataSnapshot.child("password").getValue(String.class);
+                    String confirmPassword = dataSnapshot.child("password").getValue(String.class);
 
-                        } else {
-                            Exception exception = task.getException();
-                            if (exception != null) {
-                                if (exception != null) {
-                                    Log.d("DatabaseError", "Hiba történt a törlés során: " + exception.getMessage());
-                                } else {
-                                    Log.d("DatabaseError", "Ismeretlen hiba történt a törlés során.");
-                                }
+                    profileUsernameEditText.setText(username);
+                    profileEmailEditText.setText(email);
+                    profilePasswordEditText.setText(password);
+                    profileConfirmPasswordEditText.setText(confirmPassword);
+                }
+            }
+            @Override
+            public void onCancelled(DatabaseError databaseError) {
+
+            }
+        });
+    }
+    private void deleteUserDataFromDatabase() {
+        //String userId = FirebaseAuth.getInstance().getCurrentUser().getUid();
+
+        FirebaseUser user = FirebaseAuth.getInstance().getCurrentUser();
+        if (user != null) {
+            // A felhasználó be van jelentkezve, így biztonságosan használhatjuk a FirebaseUser objektumot.
+            String userId = user.getUid();
+            DatabaseReference profileDb = FirebaseDatabase.getInstance().getReference("users").child(userId);
+            // További műveletek a felhasználóval...
+            profileDb.removeValue()
+                    .addOnCompleteListener(new OnCompleteListener<Void>() {
+                        @Override
+                        public void onComplete(@NonNull Task<Void> task) {
+                            if (task.isSuccessful()) {
+                                Toast.makeText(ProfileActivity.this, "HURRÁ törölve!", Toast.LENGTH_SHORT).show();
+                            }
+                            else {
+                                Toast.makeText(ProfileActivity.this, "NEMÁR!", Toast.LENGTH_SHORT).show();
                             }
                         }
-                    }
-                });
+                    });
+        }
     }
+
 }
